Renames util to io and adds useful comment
