+++ tests/test_csv.py
@@ -1,6 +1,6 @@
 import pytest
 from phaser import Pipeline, Phase, DataErrorException
-from phaser.util import read_csv
+from phaser.io import read_csv
 
 def test_duplicate_column_names(tmpdir):
     with open(tmpdir / 'dupe-column-name.csv', 'w') as f:
